Use expect/actual funs instead of classes
