+++ normalized-cache-incubating/src/appleMain/kotlin/com/apollographql/cache/normalized/internal/Lock.apple.kt
@@ -4,14 +4,14 @@ import kotlinx.atomicfu.locks.ReentrantLock
 import kotlinx.atomicfu.locks.reentrantLock
 import kotlinx.atomicfu.locks.withLock
 
-internal actual class Lock {
+internal actual fun Lock(): Lock = object : Lock {
   private val lock: ReentrantLock = reentrantLock()
 
-  actual fun <T> read(block: () -> T): T {
+  override fun <T> read(block: () -> T): T {
     return lock.withLock(block)
   }
 
-  actual fun <T> write(block: () -> T): T {
+  override fun <T> write(block: () -> T): T {
     return lock.withLock(block)
   }
 }
