+++ normalized-cache-incubating/src/jsCommonMain/kotlin/com/apollographql/cache/normalized/internal/Lock.jsCommon.kt
@@ -1,17 +1,11 @@
 package com.apollographql.cache.normalized.internal
 
-/**
- * A lock with read/write semantics where possible.
- *
- * - uses Java's `ReentrantReadWriteLock` on the JVM
- * - uses AtomicFu's [ReentrantLock] on Native (read and write are not distinguished)
- */
-internal actual class Lock {
-  actual fun <T> read(block: () -> T): T {
+internal actual fun Lock(): Lock = object : Lock {
+  override fun <T> read(block: () -> T): T {
     return block()
   }
 
-  actual fun <T> write(block: () -> T): T {
+  override fun <T> write(block: () -> T): T {
     return block()
   }
 }
