+++ normalized-cache-incubating/src/jvmMain/kotlin/com/apollographql/cache/normalized/internal/Lock.jvm.kt
@@ -4,16 +4,16 @@ import java.util.concurrent.locks.ReentrantReadWriteLock
 import kotlin.concurrent.read
 import kotlin.concurrent.write
 
-internal actual class Lock {
+internal actual fun Lock(): Lock = object : Lock {
   private val lock = ReentrantReadWriteLock()
 
-  actual fun <T> read(block: () -> T): T {
+  override fun <T> read(block: () -> T): T {
     return lock.read {
       block()
     }
   }
 
-  actual fun <T> write(block: () -> T): T {
+  override fun <T> write(block: () -> T): T {
     return lock.write {
       block()
     }
