+++ normalized-cache-sqlite-incubating/api/normalized-cache-sqlite-incubating.klib.api
@@ -123,12 +123,8 @@ final class com.apollographql.cache.normalized.sql/SqlNormalizedCache : com.apol
     final fun remove(com.apollographql.cache.normalized.api/CacheKey, kotlin/Boolean): kotlin/Boolean // com.apollographql.cache.normalized.sql/SqlNormalizedCache.remove|remove(com.apollographql.cache.normalized.api.CacheKey;kotlin.Boolean){}[0]
     final fun remove(kotlin/String): kotlin/Int // com.apollographql.cache.normalized.sql/SqlNormalizedCache.remove|remove(kotlin.String){}[0]
 }
-final class com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory : com.apollographql.cache.normalized.api/NormalizedCacheFactory { // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory|null[0]
-    constructor <init>() // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory.<init>|<init>(){}[0]
-    constructor <init>(app.cash.sqldelight.db/SqlDriver) // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory.<init>|<init>(app.cash.sqldelight.db.SqlDriver){}[0]
-    constructor <init>(kotlin/String? = ...) // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory.<init>|<init>(kotlin.String?){}[0]
-    constructor <init>(kotlin/String?, kotlin/String?) // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory.<init>|<init>(kotlin.String?;kotlin.String?){}[0]
-    final fun create(): com.apollographql.cache.normalized.api/NormalizedCache // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory.create|create(){}[0]
-}
 final const val com.apollographql.cache.normalized.sql/VERSION // com.apollographql.cache.normalized.sql/VERSION|{}VERSION[0]
     final fun <get-VERSION>(): kotlin/String // com.apollographql.cache.normalized.sql/VERSION.<get-VERSION>|<get-VERSION>(){}[0]
+final fun com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory(app.cash.sqldelight.db/SqlDriver): com.apollographql.cache.normalized.api/NormalizedCacheFactory // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory|SqlNormalizedCacheFactory(app.cash.sqldelight.db.SqlDriver){}[0]
+final fun com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory(kotlin/String? = ...): com.apollographql.cache.normalized.api/NormalizedCacheFactory // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory|SqlNormalizedCacheFactory(kotlin.String?){}[0]
+final fun com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory(kotlin/String?, kotlin/String?): com.apollographql.cache.normalized.api/NormalizedCacheFactory // com.apollographql.cache.normalized.sql/SqlNormalizedCacheFactory|SqlNormalizedCacheFactory(kotlin.String?;kotlin.String?){}[0]
