+++ normalized-cache-sqlite-incubating/src/appleMain/kotlin/com/apollographql/cache/normalized/sql/SqlNormalizedCacheFactory.apple.kt
@@ -1,0 +1,27 @@
+package com.apollographql.cache.normalized.sql
+
+import app.cash.sqldelight.db.SqlDriver
+import com.apollographql.cache.normalized.api.NormalizedCache
+import com.apollographql.cache.normalized.api.NormalizedCacheFactory
+import com.apollographql.cache.normalized.sql.internal.createDriver
+import com.apollographql.cache.normalized.sql.internal.createRecordDatabase
+import com.apollographql.cache.normalized.sql.internal.getSchema
+
+actual fun SqlNormalizedCacheFactory(driver: SqlDriver): NormalizedCacheFactory = object : NormalizedCacheFactory() {
+  override fun create(): NormalizedCache {
+    return SqlNormalizedCache(createRecordDatabase(driver))
+  }
+}
+
+/**
+ * @param name the name of the database or null for an in-memory database
+ * @param baseDir the baseDirectory where to store the database.
+ * [baseDir] must exist and be a directory
+ * If [baseDir] is a relative path, it will be interpreted relative to the current working directory
+ */
+fun SqlNormalizedCacheFactory(
+    name: String?,
+    baseDir: String?,
+): NormalizedCacheFactory = SqlNormalizedCacheFactory(createDriver(name, baseDir, getSchema()))
+
+actual fun SqlNormalizedCacheFactory(name: String?): NormalizedCacheFactory = SqlNormalizedCacheFactory(name, null)
